chore: move draw into its' own file
